==> 01. EDA_Detail_Incident.ipynb <==
# 01. Exploratory Data Analysis and Preliminary Cleaning


import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
from matplotlib import pyplot as plt
from pandas_profiling import ProfileReport
from pathlib import Path

df = pd.read_csv("data/Detail_Incident.csv", sep=";", decimal=',')

df.columns

Remove empty rows and columns

df.dropna(axis='columns', how='all', inplace=True)

df.dropna(axis='rows', how='all', inplace=True)

df.shape

Adjust column names for easier reference

df.columns = df.columns.str.replace(' ', '_')
df.columns = df.columns.str.replace('(', '')
df.columns = df.columns.str.replace(')', '')
df.columns = df.columns.str.replace('#', 'Count')

df.columns

Convert date columns to datetime

colsDatetime = ['Open_Time', 'Reopen_Time', 'Resolved_Time', 'Close_Time']

for i in colsDatetime:
    df[i] = pd.to_datetime(df[i], format='%d/%m/%Y %H:%M:%S', errors='coerce' )

df.dtypes

Investigate `Urgency` as an object

df.Urgency.value_counts()

Fix `Urgency`, convert it along with `Impact` and `Priority` to string

df.Impact = df.Impact.astype(str).str[:1]
df.Priority = df.Priority.astype(str).str[:1]
df.Urgency = df.Urgency.astype(str).str[:1]

df.Urgency.value_counts()

df.dtypes


Output file and create profile report

with open("data/01.a.Detail_Incident.csv",'w') as f:
    df.to_csv(f, index=False)

profile = ProfileReport(df, title="Profile of BPIC 2014 Detail_Incident Data after Initial Cleaning", html={'style': {'full_width': True}})

profile.to_file(Path(str("reports/01.b.Detail_Incident_Profile.html")))
