==> 02. Cleaning_Detail_Incident.ipynb <==
# 02. Cleaning the Source Data Set

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
from matplotlib import pyplot as plt
from pandas_profiling import ProfileReport
from pathlib import Path

df = pd.read_csv("data/01.a.Detail_Incident.csv", parse_dates=['Open_Time', 'Reopen_Time', 'Resolved_Time','Close_Time', ])

df.dtypes

## Drop Records where Resolved_Time is Missing

df.iloc[:,13:17].isnull().sum()

df = df.dropna(subset=['Resolved_Time'])

df.iloc[:,13:17].isnull().sum()

## Limit timeframe of all records

greater than 1 october 2013

less than 31 march 2014


df = df[df['Open_Time'] >= pd.to_datetime('10-01-2013')]

df.iloc[:,13:17].describe()

## Deal with Status of 'work in progress'

df.Status.value_counts()

df = df[ df['Status'] == 'Closed' ]

df.Status.value_counts()

## Remove non-incident records

print(df.Category.value_counts())

df = df[ df['Category'] == 'incident' ]

print(df.Category.value_counts())
print(df.Status.value_counts())
print(df.Alert_Status.value_counts())

## Deal with Reopen_Time Missing Values


df.Reopen_Time.isnull().sum()

df['ReopenedFlag'] = ~ df.Reopen_Time.isnull()

df['ReopenedFlag'] = df['ReopenedFlag'].astype(int)

df['ReopenedFlag'].value_counts()

## Set Missing to Zero for `Count_Related_Changes`, `Count_Related_Incidents`, and `Count_Related_Interactions` 

print(df['Count_Related_Changes'].isnull().sum())
print(df['Count_Related_Incidents'].isnull().sum())
print(df['Count_Related_Interactions'].isnull().sum())


df['Count_Related_Changes'] = df['Count_Related_Changes'].fillna(0)
df['Count_Related_Incidents'] = df['Count_Related_Incidents'].fillna(0)
df['Count_Related_Interactions'] = df['Count_Related_Interactions'].fillna(0)


print(df['Count_Related_Changes'].isnull().sum())
print(df['Count_Related_Incidents'].isnull().sum())
print(df['Count_Related_Interactions'].isnull().sum())

## Set Missing to "Not Applicable" for `Related_Change`

df['Related_Change'].value_counts().sum()

df['Related_Change'] = df['Related_Change'].fillna("Not Applicable")

df['Related_Change'].value_counts()

## Drop columns 

* with constant values,   

* longer needed (`Reopen_Time`)



df = df.drop(['Category', 'Status', 'Alert_Status', 'Reopen_Time'], axis='columns')

df.columns

## END and OUTPUT

with open("data/02.a.Detail_Incident.csv",'w') as f:
    df.to_csv(f, index=False)

df.reset_index(drop=True, inplace=True)
profile = ProfileReport(df, title="Profile of BPIC 2014 Detail_Incident Data after Secondary Cleaning", html={'style': {'full_width': True}})

profile.to_file(Path(str("reports/02.b.Detail_Incident_Profile.html")))
