==> 04. Final Data Prep.ipynb <==
# 04. Final Data Preparation

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
from matplotlib import pyplot as plt
from pandas_profiling import ProfileReport
from pathlib import Path

df = pd.read_csv("data/03.a.Detail_Incident.csv", parse_dates=['Open_Time',  'Resolved_Time','Close_Time'])

df.dtypes

df.head()

df.Priority.value_counts()

df["Priority"] = pd.cut(x=df['Priority'], bins=[-1,1,2,3,4,5], labels=["1 Very High", "2 High", "3 Medium", "4 Low", "5 Very Low"])
df["Impact"] = pd.cut(x=df['Impact'], bins=[-1,1,2,3,4,5], labels=["1 Very High", "2 High", "3 Medium", "4 Low", "5 Very Low"])
df["Urgency"] = pd.cut(x=df['Urgency'], bins=[-1,1,2,3,4,5], labels=["1 Very High", "2 High", "3 Medium", "4 Low", "5 Very Low"])



df.Priority.value_counts()

df['Open_Time_HourOfDay'] = df.Open_Time.dt.hour
df['Resolved_Time_HourOfDay'] = df.Resolved_Time.dt.hour
df['Close_Time_HourOfDay'] = df.Close_Time.dt.hour

df.head()

df['Open_Time_DayOfWeek'] = df.Open_Time.dt.day_name()
df['Resolved_Time_DayOfWeek'] = df.Resolved_Time.dt.day_name()
df['Close_Time_DayOfWeek'] = df.Close_Time.dt.day_name()

df.head()

df['OpenShift'] = pd.cut(x=df['Open_Time_HourOfDay'], bins=[-1, 8, 16, 25], labels=['Night','Day','Evening'])
df['ResolvedShift'] = pd.cut(x=df['Resolved_Time_HourOfDay'], bins=[-1, 8, 16, 25], labels=['Night','Day','Evening'])
df['CloseShift'] = pd.cut(x=df['Close_Time_HourOfDay'], bins=[-1, 8, 16, 25], labels=['Night','Day','Evening'])

df.head()

df.columns

df['CI_TypeSubType_aff'] = df.CI_Type_aff + "-" + df.CI_Subtype_aff
df['CI_TypeSubType_CBy'] = df.CI_Type_CBy + "-" + df.CI_Subtype_CBy

df.head()

df = df.drop([ "CI_Type_aff", "CI_Subtype_aff", "CI_Type_CBy", "CI_Subtype_CBy"], axis='columns')

df = df.drop(['Incident_ID', "Related_Interaction",  "Related_Change"], axis='columns')

df.columns


dfAtOpen = df[['CI_Name_aff', 'Service_Component_WBS_aff', 'Impact', 'Urgency',
        'KM_number', 'Count_Related_Interactions', 'Count_Related_Incidents',
       'Count_Related_Changes', 'SLAFail',
       'Open_Time_HourOfDay',  'Open_Time_DayOfWeek',
       'CI_TypeSubType_aff']]

dfAtOpen.columns

dfAtOpen.shape

## END and OUTPUT

with open("data/04.a.Detail_Incident.csv",'w') as f:
    df.to_csv(f, index=False)

df.reset_index(drop=True, inplace=True)
profile = ProfileReport(df, title="Profile of Final BPIC 2014 Detail Incident Data", html={'style': {'full_width': True}})

profile.to_file(Path(str("reports/04.b.Detail_Incident_Profile.html")))

with open("data/04.a.Detail_Incident_AtOpen.csv",'w') as f:
    dfAtOpen.to_csv(f, index=False)

dfAtOpen.reset_index(drop=True, inplace=True)
profile = ProfileReport(dfAtOpen, title="Profile of Final BPIC 2014 Detail Incident At Open Data", html={'style': {'full_width': True}})

profile.to_file(Path(str("reports/04.b.Detail_Incident_AtOpen_Profile.html")))
