==> 05.99 Get Metrics.ipynb <==
# 00.01 Get Metrics

Goal: identify the factors that most contribute to SLAFail

Tuning Adjustments: Focus on finding the most predictive set of predictor variables

Compare the results of the following notebooks: 







# Load libraries
import pandas as pd
import numpy as np
import pickle


data = {}

##### Transformation Metrics
########### change to code cell to run and back to markdown to 'comment'

files = {"MEstMetadata" :  "data/05.01.a BareBones MEst.pkl",
         "TargetMetadata" : "data/05.01.b BareBones Target.pkl", 
         "WOEMetadata" : "data/05.01.c BareBones WOE.pkl", 
         "HelmertMetadata" : "data/05.01.d BareBones HelmertEncoder.pkl",
         "RobustMetadata" : "data/05.01.e BareBones RobustScaler.pkl" }
         
outFilename = "reports/05.01 Transformation Metrics.csv"

##### Feature Selection Metrics

########### change to code cell to run and back to markdown to 'comment'

files = {"KBestFMetadata" :  "data/05.02.a Feature Select KBest f_classif.pkl",
         "KBestMIMetadata" : "data/05.02.b Feature Select KBest MI.pkl", 
         "RFEMetadata" : "data/05.02.c Feature Select RFE.pkl" }

outFilename = "reports/05.01 Feature Selection Metrics.csv"

for key, value in files.items():
    print(key + ": " + value)
    with open(value, 'rb') as fi:
        data[key] = pickle.load(fi)
    


df = pd.DataFrame(data)
df

with open(outFilename, 'w') as fo:
    df.to_csv(fo)
